fix: stat dims, frp marginal, flattened strs; scalar stat policy
Several inconsistencies in the new dimension structure for
statistics are fixed here, in frps and kinds.

A small but serious bug in marginalization of frps
and in the flattening of string values in kind factories.

Changed policy on statistics decorator to prefer more
expansive tuple arity when inferring from a function
with only one parameter. This forces the dim=1
when you are doing scalar specific things, which seems
like the most sensible way to do it. This is a reversion
to an old policy.