feat: in conditional_kind/frp with fn arg, codim=1 => scalar passed
With a dict arg, the conversion is already done, but it is
convenient to define conditional fn's that take a scalar,
so the codim=1 hint indicates this explicitly.  For all other
codims (not None), the sizes are checked but a tuple is passed.
If codim=None, then no checks are done and a tuple is passed.