policy: monoidal statistics return unit on Kind.empty
Applies as well to any statistic that accepts no arguments.
This is done with a special check in transform which would
otherwise be skipped since _canonical is empty.

The key use case is for Monoidal statistics like Sum,
but might be useful otherwise.

Min and Max raise the issue of infinities that break the
kind printing and have to be handled.