fix?: use as_numeric in as_quantity numeric vs. as_nice_numeric
This leads to some numerical results coming out better,
but it remains to be seen if the reason I changed to
as_nice_numeric for numeric conversion holds true.

Right now conversion is done to make nice in Kind.show_full.
Perhaps need to add an as_nice* in expectation as well
to make the values nicer.