fix: Issue 2: typing and input for conditional Kinds and FRPs
Handled common special cases to avoid type errors in
conditional_kind and conditional_frp that arises because
dict is invariant.  So I expanded the types for the
most common cases of scalar dictionaries.  There may
be a few others, but we'll see.... This is a hack of sorts,
but there is no real way around that as far as I can see.

Also improved the errors on bad input to ConditionalKind and
ConditionalFRP and made conditional_kind and conditional_frp
smarter when given a ConditionalKind or ConditionalFRP
(or Kind or FRP).