feat: manage freshness, values flow properly, is_kinded clarified
New method is_fresh, updating of cached_values down an expression
tree; automatic expression on fresh values keeps FRPs fresh
and makes the data flow properly; all the main operations
now support this. The meaning if is_kinded is clarified so
the invariant is either expr is not None or is_kinded,
but the or is non-exclusive. (That is, expressions can have
kinds which are used.)

This is a significant and breaking change