feat: added ^ operator with vec_tuples on left: v ^ f === f(v)
This should have been done from the beginning. It makes the
arrow operator more consistent and useful with factories.
I want to ensure that it causes no other problems, but I don't
see how it would as it only applies when vec tuples are on the
left, which eliminates  FRP ^ Stat, Kind ^ Stat, and Stat ^ Stat
versions of the operator.